Trait jf_commitment::CommitmentScheme

pub trait CommitmentScheme {
    type Input;
    type Output: Clone + Debug + PartialEq + Eq + Hash;
    type Randomness: Clone + Debug + PartialEq + Eq + UniformRand;
    type Error: Error;

    // Required methods
    fn commit<T: Borrow<Self::Input>>(
        input: T,
        r: Option<&Self::Randomness>,
    ) -> Result<Self::Output, Self::Error>;
    fn verify<T: Borrow<Self::Input>>(
        input: T,
        r: Option<&Self::Randomness>,
        comm: &Self::Output,
    ) -> Result<Result<(), ()>, Self::Error>;
}

Required Associated Types

type Input

Input to the commitment

type Output: Clone + Debug + PartialEq + Eq + Hash

The type of output commitment value

type Randomness: Clone + Debug + PartialEq + Eq + UniformRand

The type of the hiding/blinding factor

type Error: Error

Error type

Required Methods

fn commit<T: Borrow<Self::Input>>( input: T, r: Option<&Self::Randomness>, ) -> Result<Self::Output, Self::Error>

Commit algorithm that takes input and blinding randomness r (optional for hiding commitment schemes), outputs a commitment.

fn verify<T: Borrow<Self::Input>>( input: T, r: Option<&Self::Randomness>, comm: &Self::Output, ) -> Result<Result<(), ()>, Self::Error>

Verify algorithm that output Ok if accepted, or Err if rejected.

Object Safety

This trait is not object safe.